+++ src/main/java/com/lucasdevx/LibraryLoanSystem/controller/BookController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.lucasdevx.LibraryLoanSystem.dto.BookDTO;
+import com.lucasdevx.LibraryLoanSystem.exception.ObjectIsNullException;
 import com.lucasdevx.LibraryLoanSystem.model.Book;
 import com.lucasdevx.LibraryLoanSystem.service.BookService;
 
@@ -27,6 +28,10 @@ public class BookController {
 	
 	@PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public BookDTO insert(@RequestBody BookDTO bookDTO) {
+		if(bookDTO == null) {
+			throw new NullPointerException("Id is null");
+		}
+		
 		Book book = bookService.insert( bookService.parseToBook(bookDTO));
 		
 		return bookService.parseToBookDTO(book);
@@ -34,6 +39,10 @@ public class BookController {
 	
 	@GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
 	public BookDTO getById(@PathVariable Long id) {
+		if(id <= 0) {
+			throw new IllegalArgumentException("Invalid id");
+		}
+		
 		Book book = bookService.getById(id);
 		return bookService.parseToBookDTO(book);
 	}
@@ -50,6 +59,16 @@ public class BookController {
 	
 	@PutMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public BookDTO update(@RequestBody BookDTO bookDTO) {
+		if(bookDTO == null) {
+			throw new ObjectIsNullException();
+		}
+		if(bookDTO.id() == null) {
+			throw new NullPointerException("Id is null");
+		}
+		if(bookDTO.id() <= 0) {
+			throw new IllegalArgumentException("Invalid id");
+		}
+		
 		Book newBook = bookService.parseToBook(bookDTO);
 		Book bookUpdated = bookService.update(newBook);
 		
@@ -58,6 +77,9 @@ public class BookController {
 	
 	@DeleteMapping("/{id}")
 	public ResponseEntity<?> delete(@PathVariable Long id){
+		if(id <= 0) {
+			throw new IllegalArgumentException("Invalid id");
+		}
 		bookService.delete(id);
 		return ResponseEntity.noContent().build();
 	}
